+++ lib/addressfinder.rb
@@ -9,6 +9,7 @@ require 'addressfinder/address_info'
 require 'addressfinder/address_search'
 require 'addressfinder/address_autocomplete'
 require 'addressfinder/bulk'
+require 'addressfinder/email/verification'
 require 'addressfinder/errors'
 require 'addressfinder/util'
 require 'addressfinder/http'
@@ -59,6 +60,10 @@ module AddressFinder
 
     def address_info(args={})
       AddressFinder::AddressInfo.new(params: args, http: AddressFinder::HTTP.new(configuration)).perform.result
+    end
+
+    def email_verification(args={})
+      AddressFinder::Email::Verification.new(params: args, http: AddressFinder::HTTP.new(configuration)).perform.result
     end
 
     def bulk(&block)